requirements.txt
openai==0.22.0
mysql-connector-python==8.0.30
tweepy==4.10.0
